src/requirements.txt
Flask==2.3.2
functions-framework==3.4.0
google-cloud-logging==3.5.0
slack_bolt==1.18.0
beautifulsoup4==4.12.2
requests==2.31.0
